Refactor: create separate components for each chart
