+++ src/constants.ts
@@ -9,3 +9,11 @@ export const CHART_MARGINS = {
   left: 50,
   bottom: 50,
 };
+
+export const CHART_COLORS = [
+  '#0088FE',
+  '#00C49F',
+  '#FFBB28',
+  '#FF8042',
+  '#FF5722',
+];
